refactor(Sheet): replaced SheetContent with SheetContentFunctional
